Timestamp file now builds with the hash of all src files timestamps in the name. This means we can detect changes if files get removed.
Timestamp is stored AFTER the code is compiled so that if there are compilation errors the timestamp will not be created and it will force to build the code again. This means we need to add additional configuration to pom.xml.

Timestamp file is stored outside the target so that if we clean the target we still have the timestamp.

We detect if output folder is missing so even if timestamp file matches the source, target will be rebuild regardless.